optimize 'PlayerClient#getPlayProgress'

+++ player/src/main/java/snow/player/PlayerClient.java
@@ -484,7 +484,7 @@ public class PlayerClient implements Player {
      *
      * @return 播放进度
      */
-    public long getPlayProgress() {
+    public int getPlayProgress() {
         return mPlayerStateHolder.mPlayerState.getPlayProgress();
     }
